[refractor] seperate follow button
